misc: Link to the PDFs of this manual
Somebody ending in this repository might not know what Voyage
is. Take the description from the  manual, link to the directory
that will hold the PDFs and point to other books and booklets too.

Feel free to proposed improved wording of the introduction and
link to the PDFs. Instead of pointing to a directory I would have
preferred to point to the latest PDF but there doesn't seem to be
such a link. Better point to the directory instead of having a
dead link in the future.